Moved "Settings" out of tree and into separate popup and menu
